Make sched generic version of slurm
