+++ calkit/cli/main/core.py
@@ -57,8 +57,7 @@ from calkit.cli.new import new_app
 from calkit.cli.notebooks import notebooks_app
 from calkit.cli.office import office_app
 from calkit.cli.overleaf import overleaf_app
-from calkit.cli.pbs import pbs_app
-from calkit.cli.slurm import slurm_app
+from calkit.cli.sched import scheduler_app
 from calkit.cli.update import update_app
 
 app = typer.Typer(
@@ -82,8 +81,11 @@ app.add_typer(check_app, name="check", help="Check things.")
 app.add_typer(latex_app, name="latex|tex", help="Work with LaTeX.")
 app.add_typer(overleaf_app, name="overleaf|ol", help="Interact with Overleaf.")
 app.add_typer(cloud_app, name="cloud", help="Interact with a Calkit Cloud.")
-app.add_typer(slurm_app, name="slurm", help="Work with SLURM.")
-app.add_typer(pbs_app, name="pbs", help="Work with PBS.")
+app.add_typer(
+    scheduler_app,
+    name="sched|slurm",
+    help="Work with a job scheduler (SLURM or PBS).",
+)
 app.add_typer(dev_app, name="dev", help="Developer tools.", hidden=True)
